third_party/blink/web_tests/external/wpt/html/syntax/xmldecl/xmldecl-2.html

<!doctype html>
<meta charset="windows-1252">
<title>Bogo-XML declaration</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script>
setup({explicit_done:true});
window.onload = function() {
    let divs = document.getElementsByTagName("div");
    for (let i = 0; i < divs.length; ++i) {
        let div = divs[i];
        let expectation = div.className;
        let iframes = div.getElementsByTagName("iframe");
        for (let j = 0; j < iframes.length; ++j) {
            let iframe = iframes[j];
            let src = iframe.src;
            let doc = iframe.contentWindow.document;
            test(function() {
                assert_equals(doc.characterSet, expectation, 'Check');
                if (expectation == "windows-1251" || expectation == "windows-1252" && !(src.endsWith("/XML.htm") || src.endsWith("/XML-trail.htm"))) {
                    let fc = doc.firstChild;
                    assert_equals(fc.nodeType, Node.COMMENT_NODE, 'Should have comment node');
                    assert_true(fc.nodeValue.startsWith("?xml"), 'Should start with ?xml');
                } else if (expectation == "UTF-16BE" || expectation == "UTF-16LE") {
                    let fc = doc.firstChild;
                    assert_equals(fc.nodeType, Node.COMMENT_NODE, 'Should have comment node');
                    assert_true(fc.nodeValue.startsWith("?x"), 'Should start with ?x');
                }
            }, "Check encoding " + expectation + ", " + src.substring(src.lastIndexOf("/") + 1));
        }
    }
    done();
};
</script>

<section style="display: none;">
<div class="windows-1252">
<iframe src="support/encodingencoding.htm"></iframe>
<iframe src="support/encoding-equals-encoding.htm"></iframe>
<iframe src="support/ENCODING.htm"></iframe>
<iframe src="support/gt-between-xml-and-encoding.htm"></iframe>
<iframe src="support/kilobyte-plus-one-after.htm"></iframe>
<iframe src="support/kilobyte-plus-one-before.htm"></iframe>
<iframe src="support/no-quotes.htm"></iframe>
<iframe src="support/no-quotes-space.htm"></iframe>
<iframe src="support/one-around-label.htm"></iframe>
<iframe src="support/space-around-label.htm"></iframe>
<iframe src="support/space-before.htm"></iframe>
<iframe src="support/unmatched-quotes.htm"></iframe>
<iframe src="support/XML.htm"></iframe>
<iframe src="support/zero-around-label.htm"></iframe>
<iframe src="support/encodingencoding-trail.htm"></iframe>
<iframe src="support/encoding-equals-encoding-trail.htm"></iframe>
<iframe src="support/ENCODING-trail.htm"></iframe>
<iframe src="support/gt-between-xml-and-encoding-trail.htm"></iframe>
<iframe src="support/kilobyte-plus-one-after-trail.htm"></iframe>
<iframe src="support/kilobyte-plus-one-before-trail.htm"></iframe>
<iframe src="support/no-quotes-trail.htm"></iframe>
<iframe src="support/no-quotes-space-trail.htm"></iframe>
<iframe src="support/one-around-label-trail.htm"></iframe>
<iframe src="support/space-around-label-trail.htm"></iframe>
<iframe src="support/space-before-trail.htm"></iframe>
<iframe src="support/unmatched-quotes-trail.htm"></iframe>
<iframe src="support/XML-trail.htm"></iframe>
<iframe src="support/zero-around-label-trail.htm"></iframe>
</div>
<div class="windows-1253">
<iframe src="support/xml-and-meta.htm"></iframe>
<iframe src="support/incomplete-utf-16le-and-meta.htm"></iframe>
<iframe src="support/incomplete-utf-16be-and-meta.htm"></iframe>
<iframe src="support/xml-and-meta-trail.htm"></iframe>
<iframe src="support/incomplete-utf-16le-and-meta-trail.htm"></iframe>
<iframe src="support/incomplete-utf-16be-and-meta-trail.htm"></iframe>
</div>
<div class="UTF-16LE">
<iframe src="support/utf-16le-and-meta.htm"></iframe>
<iframe src="support/utf-16le-and-meta-trail.htm"></iframe>
</div>
<div class="UTF-16BE">
<iframe src="support/utf-16be-and-meta.htm"></iframe>
<iframe src="support/utf-16be-and-meta-trail.htm"></iframe>
</div>
<div class="UTF-8">
<iframe src="support/utf-16le-vs-http.html"></iframe>
<iframe src="support/utf-16be-vs-http.html"></iframe>
<iframe src="support/xml-vs-http.html"></iframe>
<iframe src="support/utf-16le-vs-http-trail.html"></iframe>
<iframe src="support/utf-16be-vs-http-trail.html"></iframe>
<iframe src="support/xml-vs-http-trail.html"></iframe>
<iframe src="support/ascii-decl-for-utf-16.htm"></iframe>
</div>
<div class="replacement">
<iframe src="support/replacement.htm"></iframe>
<iframe src="support/replacement-trail.htm"></iframe>
</div>
</section>
